The Final Output for Drilling!
Line 228 - if(game.input.a) game.replaceScene(menu_scene);

REPLACE THIS WITH STREAK SCENE OR HARDNESS TEST SCENE WHEN YOU GET TO
IT!

Line 19 & 20 - //game.pushScene(menu_scene);
		  game.pushScene(how_to_play_scene);

Maybe force instructions to show like this or just delete line 20 and
keep 19. BUT IN ORDER TO CALL INSTRUCTIONS ON HOW TO PLAY, JUST CALL
how_to_play_scene Scene object!

line 99 - var collectedRocks = [0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0, 0];

Don’t forget that this is what you’ll use to know which minerals to
call for your scene! Refer to line 98 for what each index is